Enabled linq2db (major code refactor)

+++ Common/Models/Graphql/Logs/PartitionLog.cs
@@ -3,5 +3,5 @@ namespace YALM.Common.Models.Graphql.Logs;
 public class PartitionLog : LogBase
 {
 	public long? Bytes { get; set; }
-	public double? UsedPercentage { get; set; }
+	public decimal? UsedPercentage { get; set; }
 }
